(MODULES-1956) Commit Generated Types / Specs
Previously to use this type, it would require having the ability to
actually build the types and generate them from the DSC MOF repos and
use the MOF parser, hoping everything was set up correctly so the end
result was built types. However that keeps it from being usable out of
the box. The code generator is meant to save time for
creation/maintenance of the types, but should not be relied on to be
used every time someone would like to work on the module. It may be
best in the generation to include a note at the top that the generated
type should not be changed, but the underlying MOF that the type is
generated from.

Without this change, it would not be possible to easily use the module
from github.