(FM-2652) Vendor Wave kit files
Add DSC Resource Wave Kit files into a vendor folder. We must start
with the ones that match what we can currently produce based on the
fact that we are using a MOF parser that doesn't work with MOF files as
produced by default from DSC Resources. Because of this, we are
vendoring the resource wave kit that the types are built on top of.
Once we have a valid replacement, we can update this folder with latest
components.