(FM-2562) Add vendored path to PSModulePath
Add the vendored path to the PSModulePath by symlinking the vendored
folder into the program files modules path where all custom modules are
vendored. Without this change, we are not able to vendor the modules and
make it easy for folks to start using the module with little more than
WMF 5 installed.

Use a different output path for the resource generated MOF. Overwriting and
deleting the system32\Configuration MOFs messes up PowerShell DSC state in
WMF5 where it may have been okay in WMF4.

Import resources with ModuleName and ModuleVersion since they should match
exactly what we are vendoring, except for the possibility of items from the
PSDesiredStateConfiguration module. We have less control over the version
that comes with the built-in resources so only import by ModuleName.