(MODULES-1960) use Invoke-DscResource directly
Invokes Test and Set on DSC Resources directly using the properties
provided by the user. Supports DSC Resource Module version resolution
for ResourceKit DSC Resources. Will properly set LCM RefreshMode to
Disabled if not set and return it to prior setting if it has to change
it. The module will add a symlink to the DSC Resource vendored path if
not present.

DSC spec type test generation adjusted to look for Invoke-DscResource
specific details.

Note: There is no error-handling around enabling / disabling the LCM,
but this will be addressed in a future commit that moves LCM changes
out of the individual resource.  There are also some issues around
properly converting from Ruby types to PowerShell types (in particular
from strings to string arrays) that will be addressed later as well.