(MODULES-1960) use invoke-dscresource directly
Invokes test and set on DSC Resources directly using the properties
provided by the user. Supports DSC Resource Module version resolution
for ResourceKit DSC Resources. Will properly set LCM RefreshMode to
Disabled if not set and return it to prior setting if it has to change
it. The module will add a symlink to the DSC Resource vendored path if
not present.

DSC spec type test generation adjusted to look for invoke-dscresource
specific details.