(MODULES-2243) Remove LCM Refreshmode Change
Removes setting the LCM RefreshMode in the Invoke DSC Resource script
and instead bails out if it is set to anything else than Disabled.

Creates a Puppet type that will set the LCM RefreshMode to one of the
supported values. User must specify this before any DSC Resource
statments in their manifest. It uses the CIM methods to changes the LCM
refreshmode instead of PowerShell cmdlets to increase execution speed
and avoid compiling MOFs to disk