(MODULES-2292) Modify Invoke-DscResource Return Values
Modify the return values of the invoke-dscresource template to reflect
the current WMF5 changes.

Modify so that any error returns an exception message instead of generic
text
Modify the 'Test' section to return InDesiredState property
Modify the 'Set' section to return true if no exception

Modify powershell.rb to fail if any error message is returned.