(MODULES-2175) Create vendored symlink with Ruby
Symlinks created the first time during a PowerShell operation don't
always allow the symlink to be used immediately by PowerShell DSC
during the same run.

Prefer instead to create the symlink natively in Ruby. This has
the advantage that PowerShell modules brought in by the symlink
should be ready to be used by the time DSC runs.

Prefer to create the link only once. If the symlink already
exists but points to the wrong location, or if a file/folder
exists where the symlink should be, ensure that the symlink is
recreated to the proper location.

When determining the path for the symlink, check the registry
without redirection to determine where the Program Files
directory is located. `C:\Program Files\WindowsPowerShell\Modules`
is the preferred location for Machine-level PowerShell Modules and
the location where other tools, e.g. PowerShell-Get, also put
modules during installation.