(MODULES-2175) Handle symlink missing target path
When the target for a symlink no longer exists, attempting to unlink
the symlink with `Puppet::FileSystem.unlink` fails with a `Permission
Denied` error. If the target path doesn't exist, fall back to moving
the existing link in the same way we would if it were a file or
directory.