(doc)(MODULES-2408) long file names known issue
When installing the dsc module, users can run into a long file name
(LFN) issue, because the paths of some of the schema files are quite
deep, and much deeper while puppet module tool is bringing down the
resources from the forge.

Document that the issue can occur, provide a place for users to see
updates toward a final fix, and provide a workaround for anyone that
may be affected by the issue.