(MODULES-2520) Read only parameters
Some parameters can not be changed and adding type documentation to them
brought us to the idea that we should instead follow the read only pattern
for these particular parameters.

The change moves these from defaultto to defining retrieve and passing back
the value. This also converts the parameters to properties and validates
that the user does not try to set the values.