(MODULES-2244) Update rake task to remove WaitFor DSC Resources
Update rake task to remove any WaitFor DSC Resources downloaded from the
upstream DSCResources repo. We cannot support these types of DSC
Resources at this time, so add a filter to remove them after we clone
the PowerShell DSCResources repo. As a last failsafe, we check after
this task for any Waitfor resources and fail the build if they are
present.