(MODULES-2689) Update Tests to Use "MODULE_VERSION" Env Var
The integration and acceptance tests will attempt to always use the latest
module from the staging Forge currently. However, this will not always pickup
the latests pre-released version of the module if final releases exist. The
"MODULE_VERSION" environment variable is set by CI when pre-releases of a
module are published to the staging Forge.