(MODULES-2520) Validate readonly params by default
If the value of the readonly parameter is not what we expect it
to be, fail validation for the parameter. Unfortunately using
properties and overriding property.retrieve did not work as intended.

Fallback to a method that does work appropriately. Validate the readonly
parameter is the same as the default value.