(MODULES-2837) Specify DSC Resource Module Path
We ship the DSC Resources in the box that we build our types on. We used
to have to symlink
$env:ProgramFiles\WindowsPowerShell\Modules\PuppetVendoredModules to the
install path of our module in order for DSC to 'find' our vendored DSC
Resources Correctly.

With these changes we can now have our DSC Resources stored inside our
module and point Invoke-DscResource to them. This also solves the
duplicate DSC Resource problem of having the same DSC Resources in the
Modules folder.

We accomplish this by resolving the exact path to the manifest file for
the DSC Resource module and passing that to Invoke-DscResource.