(MODULES-2512) Cannot ensure absent on Service DSC Resource
If a user specifies a service as absent, when it is not running, it will
not be removed. This is because the dsc resource returns false when testing for
existence instead of true even though service exists because of the
following code `return ($State -eq "Stopped" -and $svc.Status -eq
"Stopped") -or ($svc.Status -eq "Running" -and $State -eq "Running")`
Since $state is a paramter and is defautled to 'Running', if the service
is stopped this will always return false instead of true.

If a user specifies a service as absent when the service is already not
present, it will error. This happens because we invert absent to present
to determine if resource exists. t\This causes the error because testing
for present requires a Path, but not when testing it is Absent. Since we
dont provide the path there is nothing to pass to Present, so it throws
an error during the testing phase.

This updates the readme to inform the user of workarounds to these
issues.