(MODULES-2381) Document ensure vs dsc_ensure
Either `ensure` or `dsc_ensure` can be used. Document the precedence
and what happens when you specify both. Provide a recommendation to
use `dsc_ensure` due to issues regarding `ensure => absent` reporting
success while doing nothing - this is a known issue captured in
MODULES-2966 that is now documented as well.