(MODULES-2977) Fix Vendor Path Bug in Tests
The first issue is that when an exception was thrown by failing DSC commands
it was not being caught. This caused the tests to report a pass when it should
report a failure. Second issue is that "set_dsc_cred_resource" method was
using an incorrect cleanup call. This was masked by the previous issue, but
once uncovered it caused the "create_mysql_database_server.rb" integration test
to start failing.