(MODULES-3132) MOF Parser needs to support files with a BOM
The MOF parser used in the DSC module does not allow BOM (Byte Order Mark) in
the text file.  This can be introduced by third parties as this module
depends on the upstream Powershell Github repository.  This commit adds two
steps to the rake rasks.  The first is a new rake task called dsc:sanitize.
This will parse all MOF files and remove BOM markings if found.  The second is
an additional step to the build process which will scan all MOF files for BOM
markings and fail early if found.

paired-with: James Pogran <james.pogran@puppetlabs.com>