(MODULES-3119) Remove unncessary files from module build
This commit adds extra directories to the PMT Ignore file as, build, docs and tests are
not required at runtime for DSC module.  Also added *.local to ignore local development
files.  While the addition of these files will not cause any errors, they are removed
to keep the module installation as lean as possible.