(MODULES-3221) Use PSGallery Tags for DSCResources
The PowerShell DSCResources repo uses submodules for each DSCResource
that is released to the PSGallery. Each submodule points to the actual
DSCResource git repo, but the release strategy for that repo is the
latest tag. In order to ensure we are using the 'released' PSGallery
version we modify the build rake task to use the latest git tag.

In order to document what exactly is built, a new markdown file is
added that documents the name of the DSC Resource repo, the SHA of
the tag, and the tag name. This uses a gem called ruby-git, which
is a new gem to require. This is not as big of a deal as it is in
the build gem group, which means it's only pulled in during type
building.