(MODULES-3363) Support arbitrary Ensure values
Some DSC Resources use arbitrary values for `Present` or `Absent`. The
xRunbookDirectory DSC Resource in the xSCSMA module uses `Published`
and `Draft` as valid `Present` values for `Ensure`.

Modify the type building process to explicitly treat some resources as
exceptional cases by defining a custom hash for each MOF schema type.
This provides a way to identify non-standard resources easily, and a
process which will fail hard when new resources are introduced that
don't follow the model / aren't explicitly handled by the hash. It is
allowed to omit the :absent member of the hash for resource types
that are not absentable.

Move any inline special handling for enable / disable of the built-in
MOF type MSFT_WindowsOptionalFeature to this new hash, in addition to
MSFT_xRunbookDirectory.  This normalizes all exclusions under one
process.

Further remove any exception handling inside type / spec generation
templates, and move it out to the new helpers responsible for finding
and detecting the `Ensure` values -- raise exceptions in specific
cases that likely indicate a poorly defined / implemented MOF that
should be explicitly defined in the exclusion hash.

Types have been regenerated against the first commit after 90eba64,
which happens to be the first that includes the updated xSCSMA, which
will fail without these new type building changes:

https://github.com/PowerShell/DscResources/commit/e6974e102b8af1c4283d1d48fdce1395b949a0bd

As a result, this includes 2 new types:

dsc_runbookdirectory
dsc_smavariable