(MODULES-3228) Track release versions
Generate and track versions being used in a file. When running, if not
updating, use the versions being tracked to ensure compatibility.

If a tracked version is not found when building, use/record the latest
version of that released resource. This allows the ability to add new
dsc resources automatically.

To update to the latest version of each resource, pass `true` as an
argument for `update_versions` to the import rake task or delete the
yaml file tracking the release versions.