(maint) Fix DSC import directory names
Previously a few directory names had incorrect casing which caused
DSC Resource builds to fail on Ubuntu based linux computers.  This
commit fixes the case of these directories to the correct name.

This commit also adds a comment that Git 2.2.0 or higher is now
required for the DSC resource build process.

[ci skip]