(FM-5662) Reduce the DSC type tests
This reduces the amount of boilerplate type tests generated by building
the types from the dsc resources. The original intention was to cover
all the types, but as we continue to keep pace with the released DSC
Resources the number of tests increase exponentionally without adding
any extra value. This change selects specific DSC Resources in order to
keep the safety of the unit tests without the exceedingly long runtime
penalty of the redundant tests.