(maint) Fix duplicate MOF detection
We added duplicate MOF detection in order to prevent a LIFO situation in
parsing MOFs in the powershell/DSCResources repo. This rejected any DSC
Resource name that appeared more than once in the clone. However there
are some DSC Resources that contain other DSC Resources inside them that
are only referenced by that DSC Resource and are not used outside.

This fix looks to see if the duplicate is inside the same root module
folder. If it is, it is assumed to be internal and we skip it.