(MODULES-3683) Fix maxpath issues
Implementing MODULES-3683 introduced maxpath errors with the
OfficeOnlineServerDsc and SharePointDsc as they have a child folder
structure that is different than all the other DSC Resources we have
encountered so far.

OfficeOfflineServerDsc and SharePointDsc both have a three level deep
folder structure until you actually get to the DSC Resource. This is
different than all the other DSC Resources in
https://github.com/PowerShell/DscResources. If you look at their
packaging script, they account for it by only going to the actual DSC
Resource:
https://github.com/PowerShell/SharePointDsc/blob/dev/.appveyor/appveyor.psm1#L59
when they make their nuget package for the PSGallery.

This fixes this by inspecting the path of each file and looking for
{ModuleName}/Modules/{ModuleName} and removing the extra folder.