(MODULES-4014) Restrict to PowerShell v5
Restrict exectuion to powershell version 5.0 or greater, so that a
user will get a well formed error rather than 'provider not found'

It was decided to use a new version check to halt execution early if we
aren't on a powershell version of at least 5. The existing supported?
check checks for Puppet version support as well as powershell
compatibility, but purposely is a warning not an error, as it denotes
reduced functionality. We need to stop the execution becuase we can't
operate on less than 5, so a new method is used.