(MODULES-5228) Fix PowerShellManager template path
This fixes the template path to invoke_dsc_resource.ps1.erb to match the path
convention started in commit 4a65bf8f. If the templates used by the PowerShellManager
are not namespaced using folders with the module name, then when deployed to the
puppet master file collision can happen with a LIFO winning.