(maint) Fix PowerShell manifest error checking
This commit changes the psd1 manifest parsing code to check for
the existance of the key and if it's empty on the hash of attributes
from the manifest.

Previously if the hash did not have a 'ModuleVersion' because of a
failed parse, the code looked to see if the hash was empty, causing an
exception that does not contain the DSC Resource that failed parsing.