(MODULES-5590) Fix semicolon parsing in psd1
This commit fixes the regex that parses the DSC Resource module manifest
files to correctly parse the module version.

A PowerShell psd1 manifest file follows the syntax for a PowerShell
hash. It allows newlines as well as semi-colons to deliminate each item
in the hash. It is not normal syntax to use semi-colons inside a hash
that is using newlines already, but it is still valid so the regex is
adjusted to accept it.