(MODULES-4648) Prevent zombie procs from timeout
Prior to this commit if a timeout occurs while waiting for the
connection to a named pipe then a zombie process is left on the
system as the code to close or kill it is never called. This commit
ensures that such zombie processes are removed before the error is
raised.

This commit also adds an RSPEC test to validate that if we try to
create a manager with an unreasonably low timeout value, that the
manager will not have time to open the pipes server and it will
fail to create the manager and further verify that no zombie
processes are left running.

Paired-with: Bill Hurt <william.hurt@puppet.com>

Change spec namespace