(maint) Fix ignoring PSDscResources in yml
In commit 44f2bd4 we added PSDscResources to the blacklist because it is
a side-by-side release of the same DSC Resources that are built into DSC
as well as some DSC Resources included in xPSDscResources, and we can't
build all of those without overwritting all of them.

This commit fixes this attempt by including PSDscResources in the filter
that ignores certain DSC Resources when adding them to the
dsc_resource_tags.yml file.