(maint) Remove warning for MSFT_WaitFor DSC Resources
In commit a5147a6 we added a warning to remove any DSC Resources that
match the name 'MSFT_WaitFor`. This did not prevent some DSC Resources
that do perform a `WaitFor` type function from being included over time,
and now presents a problem with updating official DSC Resources (the
warning prevents us from including parts of updated DSC Resource
modules).

Since it would potentially cause unintended problems to pick and choose
which parts of a DSC Resource were included in our build process, it was
decided to remove the warning altogether and allow `WaitFor` DSC
Resources to be included, and warn the user in the READEME that these
DSC Resources may not work as expected.