(MODULES-6748) Convert file, group and environment beaker tests to beaker rspec
We are in the process of changing the acceptance and integration tests to use
beaker rspec. This is the first PR in that effort. The existing beaker tests
have been preseved and can be called with "bundle exec rake acceptance_test"
or "bundle exec rake reference_tests".

The converted rspec tests can be called locally using the following command
"bundle exec rake beaker".