(maint) Extracts type building
This commit extracts building types from the Manager class and rake
tasks to a seperate class. This allows testing in isolation and further
refactoring work (namely global variable use) to be done later.

This commit refactors the build_dsc_types method into seperate methods
that handle building the types and the specs seperately.