(maint) Standardize folder paths in build tasks
This commit standardizes the folder paths used in both dsc.rake and
manager.rb.

The dsc.rake task lead you to believe you could change most folder paths
used to build the module using rake task arguments. This was mostly
untrue, as the folder paths were hardcoded in both dsc.rake and
manager.rb. Changing the path in the rake task would not change it for
the manager.rb class, resulting in inconsistent builds. After this
commit, changing the folder paths in the rake task will propagate down
to the manager and inherited classes.