(maint) Rename global variables
This commit renames some global variables defined in the dsc namespace
in the rake tasks. Previous variable names were confusing and often
similarly named with contrary purposes. This rename moves for more
clarity in naming, if a little verbose. Proper use of these variables is
important, as misuse results in a failed built or a corrupted build that
still 'succeeds'.