(MODULES-6838) convert package tests to rspec
We are in the process of changing the acceptance and integration tests to use
beaker rspec. This is the second PR in that effort. The existing beaker tests
have been preseved and can be called with "bundle exec rake acceptance_test"
or "bundle exec rake reference_tests".

Currently our Jenkins test automation are running the unmodified tests in
tests/acceptance and tests/integration.

The converted rspec tests can be called locally using the following command
"bundle exec rake beaker" but do not run in Jenkins. Once the all the
tests have been converted to rspec we'll remove the old tests and configure
Jenkins to run the new rspec tests.