(maint) Make build tasks work on windows
This commit enables building this module on windows platforms by
removing the charlock_holmes dependency and updating all file read/write
methods to use binary methods.

The charlock_holmes gem is only used to parse psd1 manifest files, so
the removal is constrained to one location. The gem is replaced with
inline encoding enforcement to UTF8. We don't need to worry about all
the variances that charlock_holmes worries about, we only read the psd1
file to get the version number. This is also less important as we move
to PSGallery imports and deprecate git imports.

All relevant file methods were updated to use binary reads and writes to
avoid CRLF differences when writing and reading files on windows. This
setting will 'do the right thing' regardless on windows or on linux/mac.

Note, this is dependent on PR https://github.com/puppetlabs/mof/pull/5
which updates the MOF parser to handle absolute paths on windows.