(MODULES-7039) Update gitignore for release checks
The release_checks rake task calls the check::git_ignore task which runs a git
command to list files that are ignored by git.  Normally in a CI situation there
should be no files commited that are specified in the ignore.  In this case the
build/vendor files are marked to be ignored by the 'vendor/' pattern.  This
commit adds a negation to not ignore files in the build/vendor directory.