(MODULES-7063) Fix case sensitve psd1 search
This commit fixes the code that finds the psd1 manifest file for a given
DSC Resource.

A psd1 manifest is assumed to be the same name as the DSC Resource. However
on case sensitive OS, the file name could be a different case than the
DSC Resource name, i.e SqlServerDsc vs SQLServerDsc.ps1, which would
cause File.exists? to fail to find the file even though it exists.

Dir.glob is used here to find any psd1 manifest in the module_dir folder.
This is a safe assumption because by convention there can only be one
manifest file for a DSC Resource because it follows PowerShell Module
manifest conventions. However there could be cases where there are
other psd1 files for different purposes, so we filter for ones that
match the module_name exactly, lowercased.