(MAINT) Fix acceptance tests to work with UTF-8
For reasons that are still mystical to us UTF-8 bugs deailed in MODULES-2310 are no longer presenting in
acceptance smoke step of CI. Our acceptance tests around this were expecting a UTF-8 failure and are no
longer seeing that failure and therefore are not passing any more.

This patch changes the expectation of the acceptance tests to expect UTF-8 to work.