(MODULES-8389) Ensure Build Works if DSC Module Not Tagged
Prior to this commit the build task would fail out on any PowerShell
DSC module which did not have a tag in its repository which ended in
`-PSGallery` - since our last release, this is true of multiple
modules included in the DSC Resource module maintained by Microsoft.

Even if a tracked version was specified the build still failed because
it checked for a tag regardless of whether the version was tracked.

This commit updates the rake task for checkout by only searching for
versions in the module repository if no tracked version is declared.

It also updates the list of tracked versions to include those on which
the module build would fail because of missing tags.