(MODULES-8172) Handle sensitive values in credentials
Prior to this commit, when a master and agent are of mismatched
versions (master 5.x and agent 6.x, or vice versa) and a value
is marked sensitive in a nested hash (such as an MSFT_Credential)
then the agent does not know what to do with the rich data hash
because the keys are not what it expects.

This commit adds a munge helper, munge_sensitive_hash, which
updates a hash to re-inject sensitive values where they belong.

It also ensures that this munge is added to all parameters which
the builder tags as being credentials.

Finally, it adds testing to verify that this works.

This commit *does not* build the types for the addition of the
munger. This is because, per MODULES-8389, the build process
has a bug which prevents us from building the types if any
DSC modules are missing git tags in their repository, which
several of them are.

I used the code in that ticket's patch to test and verify that
the build process does work for this change.