(MODULES-8389) Fix build process
Prior to this commit the build task would fail on any PowerShell
DSC module which did not have a tag in its repository which ended
in `-PSGallery`. Since our last release, this is true of multiple
modules included in the PowerShell DSC Resource module maintained
by Microsoft.

Even if a tracked version was specified the build still failed
because it checked for a tag regardless of whether the version
was tracked.

This commit updates the rake task for checkout to instead mark
the latest version as commit hash pinned for the submodule by
Microsoft if no PSGallery tags are found. This ensures that the
module code specified matches the code that Microsoft has pinned.