(MODULES-8856) Fix redacting sensitive values
Prior to this commit the logic for redacting sensitive values broke
the PowerShell template code when a credential is specified because
it inserts a comment into a line which is not yet finished.

This commit modifies the logic to insert a comment as a block,
preventing it from breaking the template.

It also updates the redaction logic to handle redacting multiple
secrets on a line, as well as the tests.