(FM-8072) convert acceptance tests to Litmus
In this commit we have added Litmus style acceptance tests that run in parallel with
the existing Beaker tests.

To call the tests run "bundle exec rspec ./spec/litmus" with an environment variable
of TARGET_HOST set to the target you wish to exercise the tests against.

We have also resolved problems running the tests against localhost by wrapping
every call with "powershell.exe -NoProfile -NoLogo -Command ..." because Litmus
boilerplate code in spec_helper_acceptance_litmus switches to :cmd on localhost.

Adds a distelli-manifest.yml and provision.yaml to allow for pipeline runs.
However, it should be noted that since Litmus:acceptance:parallel does not work
with modules that have both beaker and litmus tests, the pipeline runs only against
the machine specified in the litums:provision task.

NOTE: This will only run on systems with Powershell 5 and above.