(MODULES-7762) State Exact PS Version Requirement
This is a documentation change to make it clear exactly what version
of PowerShell is required for the module to function. While 99% of
customers running WMF 5.0 will have a valid version, there was at least
one release that had a PowerShell version that will not support this
module and the DSC capabilities it requires.